REF: Move CLI into separate module
