+++ tests/test_check_semantic_version.py
@@ -81,60 +81,3 @@ class TestCheckSemanticVersion(unittest.TestCase):
             os.chdir(original_working_directory)
 
         self.assertEqual(logging_context.records[0].message, "No `mkver.conf` file found. Generating one instead.")
-
-    def test_main_with_matching_versions(self):
-        """Test that the correct message and error message are returned if the current version and the expected version
-        are the same.
-        """
-        with patch("check_semantic_version.check_semantic_version.get_expected_semantic_version", return_value="0.3.9"):
-            with patch("check_semantic_version.check_semantic_version.get_current_version", return_value="0.3.9"):
-                with patch("sys.stdout") as mock_stdout:
-                    with self.assertRaises(SystemExit) as e:
-                        check_semantic_version.main(["setup.py"])
-
-                        # Check that the exit code is 0.
-                        self.assertFalse(hasattr(e, "exception"))
-
-                    message = mock_stdout.method_calls[0].args[0]
-                    self.assertIn("VERSION PASSED CHECKS:", message)
-                    self.assertIn("The current version is the same as the expected semantic version: 0.3.9.", message)
-
-    def test_main_with_non_matching_versions(self):
-        """Test that the correct message and error message are returned if the current version and the expected version
-        are not the same.
-        """
-        with patch("check_semantic_version.check_semantic_version.get_expected_semantic_version", return_value="0.5.3"):
-            with patch("check_semantic_version.check_semantic_version.get_current_version", return_value="0.3.9"):
-                with patch("sys.stdout") as mock_stdout:
-                    with self.assertRaises(SystemExit) as e:
-                        check_semantic_version.main(["setup.py"])
-
-                        # Check that the exit code is 1.
-                        self.assertEqual(e.exception.code, 1)
-
-                    message = mock_stdout.method_calls[0].args[0]
-                    self.assertIn("VERSION FAILED CHECKS:", message)
-                    self.assertIn(
-                        "The current version (0.3.9) is different from the expected semantic version (0.5.3).",
-                        message,
-                    )
-
-    def test_main_with_non_matching_versions_reversed(self):
-        """Test that the correct message and error message are returned if the current version and the expected version
-        are not the same (reversed compared to the previous test).
-        """
-        with patch("check_semantic_version.check_semantic_version.get_expected_semantic_version", return_value="0.3.9"):
-            with patch("check_semantic_version.check_semantic_version.get_current_version", return_value="0.5.3"):
-                with patch("sys.stdout") as mock_stdout:
-                    with self.assertRaises(SystemExit) as e:
-                        check_semantic_version.main(["setup.py"])
-
-                        # Check that the exit code is 1.
-                        self.assertEqual(e.exception.code, 1)
-
-                    message = mock_stdout.method_calls[0].args[0]
-                    self.assertIn("VERSION FAILED CHECKS:", message)
-                    self.assertIn(
-                        "The current version (0.5.3) is different from the expected semantic version (0.3.9).",
-                        message,
-                    )
